fix tests & create utils
