+++ lib/node-document.js
@@ -37,7 +37,7 @@ function node (state, createEdge) {
     d.transform(criticalTransform, { css: state.style.bundle.buffer })
     var source = d.bundle()
 
-    pump(source, concat(sink), function (err) {
+    pump(source, concat({ encoding: 'buffer' }, sink), function (err) {
       if (err) return done(explain(err, 'Error in documentify while operating on ' + route))
       done()
     })
