+++ lib/start.js
@@ -12,10 +12,10 @@ var nanoraf = require('nanoraf')
 var Router = require('./regex-router')
 var bankai = require('../')
 
-var StartDelimiter = clr('|', 'gray')
-var EndDelimiter = clr('|', 'gray')
-var Filled = clr('█', 'gray')
-var Empty = clr('░', 'gray')
+var StartDelimiter = '|'
+var EndDelimiter = '|'
+var Filled = '█'
+var Empty = '░'
 
 var files = [
   'manifest',
